stable init
